// src/app/footer/types.ts
import { IconDefinition } from "@fortawesome/free-solid-svg-icons";


export interface data {
    newsletterText1: string;
    newsletterText2: string; 
    updatesIcon: IconDefinition,
    locationIcon: IconDefinition;
    phoneIcon: IconDefinition;
    EmailIcon: IconDefinition;
    list1?: string[];
    list2?: string[];  
    list3?: string[];
    payments: IconDefinition[];
    socialIcons: IconDefinition[];
}